LayoutTests/platform/glib/http/tests/websocket/tests/hybi/handshake-fail-by-non-ascii-status-line-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-non-ascii-status-line' failed: The server did not accept the WebSocket handshake.
Test whether WebSocket handshake fails if the server sends status line with non-ascii characters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
